09/03/2022 Changes in tabs
changes are uploaded in the tabs in the pager and how the information is displayed
