--- app/src/main/java/com/wizeline/mobilenews/ui/custom/CustomPagerData.kt
@@ -1,6 +1,0 @@
-package com.wizeline.mobilenews.ui.custom
-
-data class CustomPagerData(
-    val tabName: String,
-    val screen: Int
-)
