+++ app/src/main/java/com/wizeline/mobilenews/ui/dashboard/GlobalNewsPage.kt
@@ -1,81 +1,167 @@
 package com.wizeline.mobilenews.ui.dashboard
 
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.OutlinedTextField
-import androidx.compose.material.TextFieldDefaults
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.input.TextFieldValue
+import androidx.compose.ui.draw.drawWithCache
+import androidx.compose.ui.graphics.BlendMode
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.asFlow
-import androidx.navigation.NavController
 import androidx.paging.LoadState
 import androidx.paging.compose.collectAsLazyPagingItems
-import com.wizeline.mobilenews.ui.theme.CompletelyLight
+import coil.annotation.ExperimentalCoilApi
+import coil.compose.rememberImagePainter
+import com.google.accompanist.pager.ExperimentalPagerApi
+import com.wizeline.mobilenews.*
+import com.wizeline.mobilenews.R
+import com.wizeline.mobilenews.domain.models.Article
+import com.wizeline.mobilenews.ui.theme.Typography
+
+@OptIn(ExperimentalPagerApi::class)
+@Composable
+fun GlobalNewsScreen() {
+    val viewModel: GlobalNewsViewModel = hiltViewModel()
+    val articles = viewModel.getArticles().asFlow()
+    val lazyArticles = articles.collectAsLazyPagingItems()
+    LazyRow(modifier = Modifier.fillMaxSize(),
+        content = {
+            items(lazyArticles.itemCount) { index ->
+                lazyArticles[index]?.let { article -> CustomScrollableArticle(article) }
+            }
+            lazyArticles.apply {
+                when {
+                    loadState.refresh is
+                            LoadState.Loading -> {
+                        item { LoadingItem() }
+                    }
+                    loadState.append is
+                            LoadState.Loading -> {
+                        item { LoadingItem() }
+                        item { LoadingItem() }
+                    }
+                    loadState.refresh is
+                            LoadState.Error -> {
+                        item { LoadingItem() }
+                        item { LoadingItem() }
+                    }
+                    loadState.append is
+                            LoadState.Error -> {
+                        item { LoadingItem() }
+                        item { LoadingItem() }
+                    }
+                }
+            }
+
+        })
+}
 
 @Composable
-fun DashboardGlobalNews(navController: NavController) {
-    val viewModel: ArticleViewModel = hiltViewModel()
-    val textState = remember { mutableStateOf(TextFieldValue()) }
-    val list =
-        viewModel.getArticlesBySearch(textState.value.text).asFlow().collectAsLazyPagingItems()
-    Column {
-        ConstraintLayout(modifier = Modifier.fillMaxSize()) {
-            val (edit_text, news) = createRefs()
-            OutlinedTextField(
-                value = textState.value,
-                onValueChange = { query ->
-                    textState.value = query
-                    viewModel.getArticlesBySearch(query.text)
-                },
-                shape = RoundedCornerShape(30.dp),
-                colors = TextFieldDefaults.textFieldColors(
-                    backgroundColor = CompletelyLight
-                ),
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp)
-                    .constrainAs(edit_text) {
-                        start.linkTo(parent.start)
-                        end.linkTo(parent.end)
-                        top.linkTo(parent.top)
-                    }
-            )
-            LazyColumn(contentPadding = PaddingValues(
-                horizontal = 16.dp,
-                vertical = 8.dp
-            ),
-                modifier = Modifier.constrainAs(news) {
-                    top.linkTo(edit_text.bottom)
+fun CustomScrollableArticle(article: Article) {
+    val configuration = LocalConfiguration.current
+    val screenWidth = configuration.screenWidthDp
+    val screenHeight = configuration.screenHeightDp
+    ConstraintLayout(
+        modifier = Modifier
+            .padding(8.dp)
+            .width(screenWidth.dp)
+            .fillMaxHeight()
+            .verticalScroll(rememberScrollState()),
+    ) {
+        val (image, title, author, detail) = createRefs()
+        NewsImage(
+            article = article,
+            modifier = Modifier
+                .width(screenWidth.dp)
+                .height((screenHeight * 0.3).dp)
+                .constrainAs(image) {
                     start.linkTo(parent.start)
                     end.linkTo(parent.end)
-                },
-                content = {
-                    items(list.itemCount) { index ->
-                        list[index]?.let { ArticleListItem(article = it) }
-                    }
-                    list.apply {
-                        when {
-                            loadState.refresh is LoadState.Loading -> {
-                                item { LoadingItem() }
-                                item { LoadingItem() }
-                            }
-                            loadState.append is LoadState.Loading -> {
-                                item { LoadingItem() }
-                                item { LoadingItem() }
-                            }
-                            loadState.refresh is LoadState.Error -> {}
-                            loadState.append is LoadState.Error -> {}
-
-                        }
-                    }
-                })
-        }
+                    top.linkTo(parent.top)
+                }
+        )
+        Text(
+            text = article.title,
+            color = Color.LightGray,
+            textAlign = TextAlign.Center,
+            style = Typography.h1,
+            modifier = Modifier
+                .padding(8.dp)
+                .constrainAs(title) {
+                    start.linkTo(parent.start)
+                    end.linkTo(parent.end)
+                    top.linkTo(image.bottom)
+                }
+        )
+        Text(
+            text = article.author ?: EMPTY_STR,
+            color = Color.LightGray,
+            textAlign = TextAlign.Center,
+            style = Typography.body1,
+            modifier = Modifier
+                .padding(8.dp)
+                .constrainAs(author) {
+                    start.linkTo(parent.start)
+                    end.linkTo(parent.end)
+                    top.linkTo(title.bottom)
+                }
+        )
+        Text(
+            text = article.text,
+            color = Color.LightGray,
+            textAlign = TextAlign.Justify,
+            style = Typography.body1,
+            modifier = Modifier
+                .padding(8.dp)
+                .constrainAs(detail) {
+                    start.linkTo(parent.start)
+                    end.linkTo(parent.end)
+                    top.linkTo(author.bottom)
+                }
+        )
     }
+}
+
+@OptIn(ExperimentalCoilApi::class)
+@Composable
+fun NewsImage(article: Article, modifier: Modifier = Modifier) {
+    var articleImg = article.image
+    if (REGEX_IMG.toRegex() !in articleImg) {
+        articleImg = DEFAULT_ARTICLE_IMG
+    }
+    Image(
+        painter = rememberImagePainter(
+            data = articleImg,
+            builder = {
+                placeholder(R.drawable.ic_launcher_background) //Todo: Change placeholder to loading
+            }
+        ),
+        contentDescription = null,
+        contentScale = ContentScale.Crop,
+        modifier = modifier
+            .fillMaxWidth()
+            .drawWithCache {
+                val gradient = Brush.verticalGradient(
+                    colors = listOf(Color.Transparent, Color.Black),
+                    startY = size.height - (size.height / ARTICLE_IMG_GRADIENT_HEIGHT),
+                    endY = size.height
+                )
+                onDrawWithContent {
+                    drawContent()
+                    drawRect(gradient, blendMode = BlendMode.Multiply)
+                }
+            }
+
+    )
 }
