+++ app/src/main/java/com/wizeline/mobilenews/ui/navigation/TabData.kt
@@ -1,0 +1,9 @@
+package com.wizeline.mobilenews.ui.navigation
+
+import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.mutableStateOf
+
+data class TabData(
+    val tabName: String,
+    val isSelected: MutableState<Boolean> = mutableStateOf(false)
+)
