+++ app/src/main/res/values/dimens.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <dimen name="default_padding">10dp</dimen>
+    <dimen name="default_padding">5dp</dimen>
     <dimen name="padding_30">30dp</dimen>
     <dimen name="spacer_height">30dp</dimen>
     <dimen name="search_input_padding">16dp</dimen>
